converted type from i32 to f32 to better work with webgl

+++ geometry/src/lib.rs
@@ -7,24 +7,24 @@ mod macros;
 use shapes::rectangle::Rectangle;
 use shapes::triangle::Triangle;
 
-pub type Point = euclid::Point2D<i32, i32>;
+pub type Point = euclid::Point2D<f32, f32>;
 pub type Points = Vec<Point>;
-pub type Size = euclid::Size2D<i32, i32>;
+pub type Size = euclid::Size2D<f32, f32>;
 
 #[derive(Debug, Clone)]
 pub struct BoundingBox {
-    pub min_x: i32,
-    pub max_x: i32,
-    pub min_y: i32,
-    pub max_y: i32,
-    pub width: i32,
-    pub height: i32
+    pub min_x: f32,
+    pub max_x: f32,
+    pub min_y: f32,
+    pub max_y: f32,
+    pub width: f32,
+    pub height: f32
 }
 
-pub fn rectangle(name: String, x: i32, y: i32, width: i32, height: i32) -> Rectangle {
+pub fn rectangle(name: String, x: f32, y: f32, width: f32, height: f32) -> Rectangle {
     return Rectangle::new(name,Point::new(x, y), Size::new(width, height));
 }
 
-pub fn triangle(name: String, x: i32, y: i32, width: i32, height: i32) -> Triangle {
+pub fn triangle(name: String, x: f32, y: f32, width: f32, height: f32) -> Triangle {
     return Triangle::new(name,Point::new(x, y), Size::new(width, height));
 }
